+++ src/components/Layout/Layout.jsx
@@ -1,14 +1,14 @@
-import React, { useContext } from 'react'
-import ProtectedHeader from './ProtectedHeader'
-import PublicHeader from './PublicHeader'
-import { Outlet } from 'react-router-dom'
-import Container from '../UI/Container'
-import styled from 'styled-components'
-import EntireContext from '../../store/Context/EntireContext'
+import { useContext } from "react";
+import ProtectedHeader from "./ProtectedHeader";
+import PublicHeader from "./PublicHeader";
+import { Outlet } from "react-router-dom";
+import Container from "../UI/Container";
+import styled from "styled-components";
+import EntireContext from "../../store/Context/EntireContext";
 
 const Main = styled.main`
   padding: 0 1.5rem;
-`
+`;
 
 const Layout = () => {
   const { isSignIn } = useContext(EntireContext);
@@ -20,7 +20,7 @@ const Layout = () => {
         <Outlet />
       </Main>
     </Container>
-  )
-}
+  );
+};
 
-export default Layout
+export default Layout;
