+++ src/pages/public/SignIn.jsx
@@ -1,33 +1,17 @@
-import { useEffect, useState } from "react";
+import {  useEffect, useState } from "react";
 import SignInputs from "../../components/SignInputs";
-import supabase from "../../supabaseClient";
 import { useLocation, useNavigate } from "react-router-dom";
+import useSignHandler from "../../hooks/useSignHandler";
 
 const SignIn = () => {
   const navigate = useNavigate();
+  const { onSignInHandler } = useSignHandler();
   const [signInInputs, setInInputs] = useState({ email: "", password: "" });
 
   const signUpInfo = useLocation().state ?? { email: "", password: "" };
   useEffect(() => {
     signUpInfo && setInInputs({ email: signUpInfo.email, password: signUpInfo.password });
   }, []);
-
-  const onSignInHandler = async (event) => {
-    event.preventDefault();
-    const { data } = await supabase.auth.signInWithPassword({
-      email: signInInputs.email,
-      password: signInInputs.password
-    });
-    
-    if (data.user) {
-      window.localStorage.setItem("signIn", true);
-      alert("로그인이 완료되었습니다.");
-      navigate("/");
-      window.location.reload();
-    } else {
-      alert("로그인에 실패했습니다.");
-    }
-  };
 
   // const signInWithGithub = async (event) => {
   //   event.preventDefault();
@@ -54,7 +38,7 @@ const SignIn = () => {
           return <SignInputs key={index} inputs={signInInputs} setInputs={setInInputs} element={element} />;
         })}
       </fieldset>
-      <button onClick={(e) => onSignInHandler(e)}>로그인</button>
+      <button onClick={(e) => onSignInHandler(e, signInInputs)}>로그인</button>
       {/* <button onClick={(e) => signInWithGithub(e)}>github 로그인</button> */}
       <button onClick={() => navigate("/sign-up")}>회원가입</button>
     </form>
