+++ src/store/reducers/signInReducer.js
@@ -1,4 +1,4 @@
-export const signInitialState = true;
+export const signInitialState = false;
 
 export const signInReducer = (state = signInitialState, action) => {
   switch (action.type) {
